requirements.txt
boto3==1.17.17
botocore==1.20.17
Flask==1.1.2
PyNaCl==1.4.0
